chore: remove stale binaries/tests and fix false scanner comments
Clean up leftover cruft from earlier migrations and correct scanner
skip-reason comments that no longer match reality.

- Remove committed ELF binaries go-functions/list (27MB) and
  go-functions/update (31MB); go-functions/bin/ is already gitignored.
- Remove unreferenced fallacy.png (3.9MB) from the repo root. The copies
  under frontend/public-astro/public/ and frontend/admin/public/ are
  actively used and are left untouched.
- Remove go-functions/tests/parity/, which compared Go responses against
  Node.js/Rust implementations that were already deleted. The generic
  JSON structural-diff helpers it also contained (CompareJSONStructure,
  FormatDiff) are still used by tests/contract/contract_test.go, so that
  logic was extracted into tests/contract/structural_diff.go before the
  parity package was removed, keeping contract tests green.
- Remove frontend/public-astro/src/lib/performanceValidation.test.ts, an
  orphaned test with no corresponding source file.
- Remove committed spec backups: .kiro/specs/serverless-blog-aws/
  design.md.backup and design.md.backup.20251103_180234.
- Rewrite .husky/scripts/pre-commit-tests.sh:
  - Drop the infrastructure/ (CDK) branch; git ls-files infrastructure
    returns zero tracked files, so the branch was dead.
  - Unify npm/npx invocations to bun, per docs/ai-shared-rules.md.
  - Drop the frontend/public branch as well. Issue #481 is deleting the
    frontend/public directory in a parallel PR; keeping both PRs off this
    shared script avoids a guaranteed merge conflict. frontend/public-astro
    and frontend/admin branches are unchanged.
- Correct stale/false skip-reason comments in terraform/.checkov.yaml and
  terraform/.trivyignore (verified against the actual Terraform code):
  - CKV_AWS_115: dropped the false "budget alerts configured" claim (no
    aws_budgets_* resource exists in this repo); kept the real mitigation
    (CloudWatch alarm on Lambda throttles).
  - CKV2_AWS_42: updated from "planned for Phase 2" to reflect that
    aws_cloudfront_response_headers_policy.security_headers is implemented
    (terraform/modules/cdn/main.tf:466-497) and attached to every cache
    behavior.
  - CKV_AWS_76 / AVD-AWS-0001: updated from "TODO: implement access
    logging" to reflect that it is implemented in modules/api/main.tf,
    gated on local.is_production (enabled in prd, disabled in dev).

- Document a local-only golangci-lint false-positive pitfall in
  docs/ai-shared-rules.md (Environment quirks): with GOTOOLCHAIN=auto, a
  system Go newer than go.mod's requirement is reused instead of
  downloading CI's pinned 1.25.12, and golangci-lint inherits that
  mismatch when it shells out to `go` while loading packages. This
  produced a local-only goconst finding during this change's own
  verification that never appeared in CI; documented so it isn't
  mistaken for real debt next time.

Closes #482
