+++ docs/ai-shared-rules.md
@@ -30,6 +30,14 @@ Run `bun run verify` before reporting a code change as done.
 - CI runs `typecheck` on every PR (no label gate). The vitest suites for `frontend/admin`,
   `frontend/public` and `frontend/public-astro` run only on PRs carrying the `frontend`
   label, so a PR outside those paths still needs `bun run verify` locally.
+- CI's `go-lint` job pins Go to 1.25.12 (`GO_VERSION` in `.github/workflows/ci.yml`).
+  With `GOTOOLCHAIN=auto`, if the local system `go` is newer than go.mod's requirement
+  (e.g. 1.26.x), Go reuses it instead of downloading 1.25.12. `golangci-lint` shells out
+  to whatever `go` is on `PATH` while loading packages, so this skew can produce findings
+  locally that never show up in CI (seen in practice with `goconst`). If `make lint` or
+  the pre-commit hook fails only locally, suspect the Go version before touching code —
+  pin `PATH` to Go 1.25.12 and re-run; if CI is green, don't rewrite the code to satisfy
+  a local-only finding.
 
 ## Repository etiquette
