--- frontend/public-astro/src/lib/performanceValidation.test.ts
@@ -1,289 +1,0 @@
-/**
- * performanceValidation.test.ts - Integration test for build validation
- *
- * Task 7.1: パフォーマンス検証 (TDD)
- *
- * These tests verify that the actual Astro build output meets performance requirements.
- * They are designed to be run as integration tests after building.
- *
- * IMPORTANT: These tests are SKIPPED by default in regular test runs.
- * Run with: bun run test:integration to validate actual build output.
- *
- * Requirements:
- * - 11.1: Lighthouse Performance スコア 95以上
- * - 11.2: TTFB 100ms未満
- * - 11.3: FCP 1秒未満
- * - 11.4: JavaScript バンドルサイズ 50KB未満（gzip）
- * - 11.5: インタラクティブ要素のないページでJS 0生成
- * - 11.6: Brotli/Gzip圧縮対応
- */
-
-import { describe, it, expect, beforeAll } from 'vitest';
-import * as fs from 'fs';
-import * as path from 'path';
-import { gzipSync } from 'zlib';
-import {
-  type BuildArtifact,
-  DEFAULT_PERFORMANCE_BUDGET,
-  validatePerformanceBudget,
-  isCompressible,
-  formatBytes,
-} from './performanceUtils';
-
-const DIST_DIR = path.join(process.cwd(), 'dist');
-
-// Check if running in integration test mode
-const isIntegrationTest = process.env.VITEST_INTEGRATION === 'true';
-
-/**
- * Recursively get all files in a directory
- */
-function getAllFiles(dir: string, files: string[] = []): string[] {
-  if (!fs.existsSync(dir)) return files;
-
-  const entries = fs.readdirSync(dir, { withFileTypes: true });
-
-  for (const entry of entries) {
-    const fullPath = path.join(dir, entry.name);
-    if (entry.isDirectory()) {
-      getAllFiles(fullPath, files);
-    } else if (entry.isFile()) {
-      files.push(fullPath);
-    }
-  }
-
-  return files;
-}
-
-/**
- * Calculate gzipped size of a file
- */
-function getGzippedSize(filePath: string): number {
-  const content = fs.readFileSync(filePath);
-  const gzipped = gzipSync(new Uint8Array(content));
-  return gzipped.length;
-}
-
-describe('Build Output Performance Validation', () => {
-  let artifacts: BuildArtifact[];
-  let distExists: boolean;
-
-  beforeAll(() => {
-    distExists = fs.existsSync(DIST_DIR);
-
-    if (distExists) {
-      const allFiles = getAllFiles(DIST_DIR);
-      artifacts = allFiles.map((filePath) => ({
-        path: path.relative(DIST_DIR, filePath),
-        sizeBytes: fs.statSync(filePath).size,
-      }));
-    } else {
-      artifacts = [];
-    }
-  });
-
-  // Helper to skip tests when not in integration mode or no dist exists
-  const skipIfNotIntegration = () => {
-    if (!isIntegrationTest) {
-      console.log(
-        'ℹ️  Skipping: Run with bun run test:integration for build validation'
-      );
-      return true;
-    }
-    if (!distExists) {
-      console.log(
-        'ℹ️  Skipping: dist directory not found (run "bun run build" first)'
-      );
-      return true;
-    }
-    return false;
-  };
-
-  describe('Build Output Existence', () => {
-    it('should have dist directory (run "bun run build" first)', () => {
-      if (!isIntegrationTest) return; // Skip in regular test runs
-      expect(distExists).toBe(true);
-    });
-
-    it('should have at least one file in dist', () => {
-      if (skipIfNotIntegration()) return;
-      expect(artifacts.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe('Required Pages', () => {
-    it('should have index.html', () => {
-      if (skipIfNotIntegration()) return;
-      const indexFile = artifacts.find((a) => a.path === 'index.html');
-      expect(indexFile).toBeDefined();
-    });
-
-    it('should have 404.html', () => {
-      if (skipIfNotIntegration()) return;
-      const errorPage = artifacts.find((a) => a.path === '404.html');
-      expect(errorPage).toBeDefined();
-    });
-
-    it('should have about/index.html', () => {
-      if (skipIfNotIntegration()) return;
-      const aboutPage = artifacts.find(
-        (a) => a.path === 'about/index.html' || a.path === 'about\\index.html'
-      );
-      expect(aboutPage).toBeDefined();
-    });
-
-    it('should have sitemap-index.xml', () => {
-      if (skipIfNotIntegration()) return;
-      const sitemap = artifacts.find((a) => a.path === 'sitemap-index.xml');
-      expect(sitemap).toBeDefined();
-    });
-
-    it('should have rss.xml (only if API available during build)', () => {
-      if (skipIfNotIntegration()) return;
-      // RSS is only generated when API is available during build
-      // So this is a soft check - report but don't fail
-      const rss = artifacts.find((a) => a.path === 'rss.xml');
-      if (!rss) {
-        console.log(
-          'ℹ️  Note: rss.xml not found (API may not have been available during build)'
-        );
-      }
-    });
-  });
-
-  describe('JavaScript Bundle Size (Requirement 11.4)', () => {
-    it('should have total JS bundle under 50KB (gzip)', () => {
-      if (skipIfNotIntegration()) return;
-
-      const jsFiles = artifacts.filter(
-        (a) => a.path.endsWith('.js') || a.path.endsWith('.mjs')
-      );
-
-      let totalGzippedSize = 0;
-      for (const artifact of jsFiles) {
-        const filePath = path.join(DIST_DIR, artifact.path);
-        totalGzippedSize += getGzippedSize(filePath);
-      }
-
-      const budget = DEFAULT_PERFORMANCE_BUDGET.maxJsBundleSizeBytes;
-
-      // Report the current state
-      console.log(
-        `   Total JS (gzip): ${formatBytes(totalGzippedSize)} / ${formatBytes(budget)}`
-      );
-
-      expect(totalGzippedSize).toBeLessThanOrEqual(budget);
-    });
-
-    it('should report JS size for static pages (Requirement 11.5)', () => {
-      if (skipIfNotIntegration()) return;
-
-      const result = validatePerformanceBudget(
-        artifacts,
-        DEFAULT_PERFORMANCE_BUDGET
-      );
-
-      // Report JS usage
-      if (result.hasInteractiveJs) {
-        console.log(
-          `   Interactive JS detected: ${formatBytes(result.totalJsSize)} (raw)`
-        );
-        console.log('   Note: For pure SSG, consider removing client-side JS');
-      } else {
-        console.log('   ✅ Zero JS for static pages');
-      }
-
-      // Just validate raw size doesn't exceed budget
-      expect(result.totalJsSize).toBeLessThanOrEqual(
-        DEFAULT_PERFORMANCE_BUDGET.maxJsBundleSizeBytes * 4 // Allow 4x raw size since we check gzip separately
-      );
-    });
-  });
-
-  describe('Total Static Size (Requirement 6.8)', () => {
-    it('should have total static size under 50MB', () => {
-      if (skipIfNotIntegration()) return;
-
-      const result = validatePerformanceBudget(
-        artifacts,
-        DEFAULT_PERFORMANCE_BUDGET
-      );
-
-      console.log(
-        `   Total size: ${formatBytes(result.totalSize)} / ${formatBytes(DEFAULT_PERFORMANCE_BUDGET.maxTotalStaticSizeBytes)}`
-      );
-
-      expect(result.totalSize).toBeLessThanOrEqual(
-        DEFAULT_PERFORMANCE_BUDGET.maxTotalStaticSizeBytes
-      );
-    });
-  });
-
-  describe('Compressible Files (Requirement 11.6)', () => {
-    it('should have HTML files that are compressible', () => {
-      if (skipIfNotIntegration()) return;
-
-      const htmlFiles = artifacts.filter((a) => a.path.endsWith('.html'));
-
-      for (const html of htmlFiles) {
-        expect(isCompressible(html.path)).toBe(true);
-      }
-    });
-
-    it('should have CSS files that are compressible', () => {
-      if (skipIfNotIntegration()) return;
-
-      const cssFiles = artifacts.filter((a) => a.path.endsWith('.css'));
-
-      for (const css of cssFiles) {
-        expect(isCompressible(css.path)).toBe(true);
-      }
-    });
-
-    it('should have XML files that are compressible', () => {
-      if (skipIfNotIntegration()) return;
-
-      const xmlFiles = artifacts.filter((a) => a.path.endsWith('.xml'));
-
-      for (const xml of xmlFiles) {
-        expect(isCompressible(xml.path)).toBe(true);
-      }
-    });
-  });
-
-  describe('Overall Validation', () => {
-    it('should report overall performance budget status', () => {
-      if (skipIfNotIntegration()) return;
-
-      const result = validatePerformanceBudget(
-        artifacts,
-        DEFAULT_PERFORMANCE_BUDGET
-      );
-
-      console.log(`\n   Performance Budget Summary:`);
-      console.log(`   - Total files: ${artifacts.length}`);
-      console.log(`   - Total size: ${formatBytes(result.totalSize)}`);
-      console.log(`   - JS size (raw): ${formatBytes(result.totalJsSize)}`);
-      console.log(`   - CSS size: ${formatBytes(result.totalCssSize)}`);
-      console.log(`   - Has interactive JS: ${result.hasInteractiveJs}`);
-      console.log(`   - Compressible files: ${result.compressibleFilesCount}`);
-
-      if (result.errors.length > 0) {
-        console.log(`\n   ❌ Errors:`);
-        for (const error of result.errors) {
-          console.log(`      - ${error}`);
-        }
-      }
-
-      if (result.warnings.length > 0) {
-        console.log(`\n   ⚠️  Warnings:`);
-        for (const warning of result.warnings) {
-          console.log(`      - ${warning}`);
-        }
-      }
-
-      // This is an informational test - actual validation happens in other tests
-      expect(true).toBe(true);
-    });
-  });
-});
