--- go-functions/tests/parity/posts_parity_test.go
@@ -1,443 +1,0 @@
-// Package parity provides API parity tests for Posts domain functions.
-//
-// Requirements: 7.2, 7.3, 7.5
-//   - Go実装とNode.js/Rust実装のレスポンス比較
-//   - HTTPステータスコード検証
-//   - レスポンスボディ構造検証
-//   - エラーメッセージ形式検証
-package parity
-
-import (
-	"encoding/json"
-	"testing"
-
-	"github.com/aws/aws-lambda-go/events"
-
-	"serverless-blog/go-functions/internal/domain"
-)
-
-// TestCreatePost_ResponseParity tests parity between Go and Node.js CreatePost implementations
-func TestCreatePost_ResponseParity(t *testing.T) {
-	t.Run("success response should have same structure", func(t *testing.T) {
-		// Expected structure from Node.js implementation
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 201,
-			Headers: map[string]string{
-				"Content-Type":                 "application/json",
-				"Access-Control-Allow-Origin":  "*",
-				"Access-Control-Allow-Headers": "Content-Type,Authorization",
-				"Access-Control-Allow-Methods": "GET,POST,PUT,DELETE,OPTIONS",
-			},
-			Body: `{
-				"id": "uuid-123",
-				"title": "Test Post",
-				"contentMarkdown": "# Hello",
-				"contentHtml": "<h1>Hello</h1>",
-				"category": "tech",
-				"tags": ["go", "aws"],
-				"publishStatus": "draft",
-				"authorId": "author-123",
-				"createdAt": "2024-01-01T00:00:00Z",
-				"updatedAt": "2024-01-01T00:00:00Z",
-				"imageUrls": []
-			}`,
-		}
-
-		// Go implementation response structure
-		goPost := domain.BlogPost{
-			ID:              "uuid-456",
-			Title:           "Test Post",
-			ContentMarkdown: "# Hello",
-			ContentHTML:     "<h1>Hello</h1>",
-			Category:        "tech",
-			Tags:            []string{"go", "aws"},
-			PublishStatus:   "draft",
-			AuthorID:        "author-456",
-			CreatedAt:       "2024-01-01T00:00:01Z",
-			UpdatedAt:       "2024-01-01T00:00:01Z",
-			ImageURLs:       []string{},
-		}
-		goBody, _ := json.Marshal(goPost)
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 201,
-			Headers:    map[string]string{"Content-Type": "application/json"},
-			Body:       string(goBody),
-		}
-
-		// Compare with ignore dynamic fields
-		result := CompareResponsesWithOptions(goResponse, nodeResponse, CompareOptions{
-			IgnoreFields: []string{"id", "authorId", "createdAt", "updatedAt"},
-		})
-
-		if result.HasDiff {
-			t.Errorf("CreatePost response structure mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("validation error should have same format", func(t *testing.T) {
-		// Node.js validation error format
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 400,
-			Body:       `{"message":"title is required"}`,
-		}
-
-		// Go validation error format
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 400,
-			Body:       `{"message":"title is required"}`,
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("CreatePost validation error format mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("unauthorized error should have same format", func(t *testing.T) {
-		// Node.js unauthorized error
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 401,
-			Body:       `{"message":"Unauthorized"}`,
-		}
-
-		// Go unauthorized error
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 401,
-			Body:       `{"message":"unauthorized"}`,
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		// Expected to have difference in case
-		if !result.HasDiff {
-			// Note: This test documents the known difference in error message casing
-			t.Log("Note: Error message casing differs between implementations")
-		}
-	})
-}
-
-// TestGetPost_ResponseParity tests parity between Go and Node.js GetPost implementations
-func TestGetPost_ResponseParity(t *testing.T) {
-	t.Run("success response should have same structure", func(t *testing.T) {
-		// Node.js response
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body: `{
-				"id": "uuid-123",
-				"title": "Existing Post",
-				"contentMarkdown": "# Content",
-				"contentHtml": "<h1>Content</h1>",
-				"category": "tech",
-				"tags": ["test"],
-				"publishStatus": "published",
-				"authorId": "author-123",
-				"createdAt": "2024-01-01T00:00:00Z",
-				"updatedAt": "2024-01-01T00:00:00Z",
-				"publishedAt": "2024-01-01T00:00:00Z",
-				"imageUrls": []
-			}`,
-		}
-
-		// Go response
-		now := "2024-01-01T00:00:00Z"
-		goPost := domain.BlogPost{
-			ID:              "uuid-123",
-			Title:           "Existing Post",
-			ContentMarkdown: "# Content",
-			ContentHTML:     "<h1>Content</h1>",
-			Category:        "tech",
-			Tags:            []string{"test"},
-			PublishStatus:   "published",
-			AuthorID:        "author-123",
-			CreatedAt:       now,
-			UpdatedAt:       now,
-			PublishedAt:     &now,
-			ImageURLs:       []string{},
-		}
-		goBody, _ := json.Marshal(goPost)
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body:       string(goBody),
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("GetPost response structure mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("not found error should have same format", func(t *testing.T) {
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("GetPost not found error format mismatch:\n%s", result.String())
-		}
-	})
-}
-
-// TestGetPublicPost_ResponseParity tests parity for public post retrieval
-func TestGetPublicPost_ResponseParity(t *testing.T) {
-	t.Run("success response for published post", func(t *testing.T) {
-		now := "2024-01-01T00:00:00Z"
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body: `{
-				"id": "uuid-123",
-				"title": "Public Post",
-				"contentMarkdown": "# Public",
-				"contentHtml": "<h1>Public</h1>",
-				"category": "tech",
-				"tags": [],
-				"publishStatus": "published",
-				"authorId": "author-123",
-				"createdAt": "2024-01-01T00:00:00Z",
-				"updatedAt": "2024-01-01T00:00:00Z",
-				"publishedAt": "2024-01-01T00:00:00Z",
-				"imageUrls": []
-			}`,
-		}
-
-		goPost := domain.BlogPost{
-			ID:              "uuid-123",
-			Title:           "Public Post",
-			ContentMarkdown: "# Public",
-			ContentHTML:     "<h1>Public</h1>",
-			Category:        "tech",
-			Tags:            []string{},
-			PublishStatus:   "published",
-			AuthorID:        "author-123",
-			CreatedAt:       now,
-			UpdatedAt:       now,
-			PublishedAt:     &now,
-			ImageURLs:       []string{},
-		}
-		goBody, _ := json.Marshal(goPost)
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body:       string(goBody),
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("GetPublicPost response structure mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("unpublished post should return 404", func(t *testing.T) {
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("GetPublicPost 404 response mismatch:\n%s", result.String())
-		}
-	})
-}
-
-// TestListPosts_ResponseParity tests parity for list posts
-func TestListPosts_ResponseParity(t *testing.T) {
-	t.Run("success response with pagination", func(t *testing.T) {
-		nextToken := "token123"
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body: `{
-				"items": [
-					{
-						"id": "uuid-1",
-						"title": "Post 1",
-						"contentMarkdown": "# Post 1",
-						"contentHtml": "<h1>Post 1</h1>",
-						"category": "tech",
-						"tags": [],
-						"publishStatus": "published",
-						"authorId": "author-1",
-						"createdAt": "2024-01-01T00:00:00Z",
-						"updatedAt": "2024-01-01T00:00:00Z",
-						"imageUrls": []
-					}
-				],
-				"nextToken": "token123"
-			}`,
-		}
-
-		goResp := domain.ListPostsResponse{
-			Items: []domain.BlogPost{
-				{
-					ID:              "uuid-1",
-					Title:           "Post 1",
-					ContentMarkdown: "# Post 1",
-					ContentHTML:     "<h1>Post 1</h1>",
-					Category:        "tech",
-					Tags:            []string{},
-					PublishStatus:   "published",
-					AuthorID:        "author-1",
-					CreatedAt:       "2024-01-01T00:00:00Z",
-					UpdatedAt:       "2024-01-01T00:00:00Z",
-					ImageURLs:       []string{},
-				},
-			},
-			NextToken: &nextToken,
-		}
-		goBody, _ := json.Marshal(goResp)
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body:       string(goBody),
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("ListPosts response structure mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("empty list response", func(t *testing.T) {
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body:       `{"items":[]}`,
-		}
-
-		goResp := domain.ListPostsResponse{
-			Items: []domain.BlogPost{},
-		}
-		goBody, _ := json.Marshal(goResp)
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body:       string(goBody),
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("ListPosts empty response mismatch:\n%s", result.String())
-		}
-	})
-}
-
-// TestUpdatePost_ResponseParity tests parity for update posts
-func TestUpdatePost_ResponseParity(t *testing.T) {
-	t.Run("success response should have same structure", func(t *testing.T) {
-		now := "2024-01-01T00:00:00Z"
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body: `{
-				"id": "uuid-123",
-				"title": "Updated Post",
-				"contentMarkdown": "# Updated",
-				"contentHtml": "<h1>Updated</h1>",
-				"category": "tech",
-				"tags": ["updated"],
-				"publishStatus": "published",
-				"authorId": "author-123",
-				"createdAt": "2024-01-01T00:00:00Z",
-				"updatedAt": "2024-01-01T00:00:00Z",
-				"publishedAt": "2024-01-01T00:00:00Z",
-				"imageUrls": []
-			}`,
-		}
-
-		goPost := domain.BlogPost{
-			ID:              "uuid-123",
-			Title:           "Updated Post",
-			ContentMarkdown: "# Updated",
-			ContentHTML:     "<h1>Updated</h1>",
-			Category:        "tech",
-			Tags:            []string{"updated"},
-			PublishStatus:   "published",
-			AuthorID:        "author-123",
-			CreatedAt:       now,
-			UpdatedAt:       now,
-			PublishedAt:     &now,
-			ImageURLs:       []string{},
-		}
-		goBody, _ := json.Marshal(goPost)
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 200,
-			Body:       string(goBody),
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("UpdatePost response structure mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("not found error should have same format", func(t *testing.T) {
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("UpdatePost not found error mismatch:\n%s", result.String())
-		}
-	})
-}
-
-// TestDeletePost_ResponseParity tests parity for delete posts
-func TestDeletePost_ResponseParity(t *testing.T) {
-	t.Run("success should return 204 with no body", func(t *testing.T) {
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 204,
-			Body:       "",
-		}
-
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 204,
-			Body:       "",
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("DeletePost 204 response mismatch:\n%s", result.String())
-		}
-	})
-
-	t.Run("not found error should have same format", func(t *testing.T) {
-		nodeResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		goResponse := events.APIGatewayProxyResponse{
-			StatusCode: 404,
-			Body:       `{"message":"post not found"}`,
-		}
-
-		result := CompareResponses(goResponse, nodeResponse)
-
-		if result.HasDiff {
-			t.Errorf("DeletePost not found error mismatch:\n%s", result.String())
-		}
-	})
-}
